+++ src/main/java/com/table/hotpack/config/form/FormLoginSuccessHandler.java
@@ -29,7 +29,7 @@ public class FormLoginSuccessHandler extends SimpleUrlAuthenticationSuccessHandl
     public static final String REFRESH_TOKEN_COOKIE_NAME = "refresh_token";
     public static final Duration REFRESH_TOKEN_DURATION = Duration.ofDays(14);
     public static final Duration ACCESS_TOKEN_DURATION = Duration.ofDays(1);
-    public static final String REDIRECT_PATH = "/articles";
+    public static final String REDIRECT_PATH = "/";
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
