+++ src/main/resources/templates/areaList.html
@@ -66,6 +66,6 @@
 <footer>
   <p>© 2024 HOTPACK. All rights reserved.</p>
 </footer>
-<script src="/js/areaList.js"></script>
+<!--<script src="/js/areaList.js"></script>-->
 </body>
 </html>
