+++ src/main/resources/templates/main.html
@@ -1,13 +1,29 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>여행 계획</title>
+    <link rel="stylesheet" href="css/style.css">
+</head>
+<body>
 <header>
     <div class="logo">
         <img src="/img/logo.png" alt="로고">
     </div>
-    <h2>HOTPACK</h2>
-    <div id="auth-buttons">
-        <!-- JavaScript로 내용 변경 -->
+    <h2 style="color: black; font-size: 24px; text-align: center;">HOTPACK</h2>
+    <div id="auth-buttons" class="auth-buttons">
+        <!-- JavaScript가 동적으로 내용 변경 -->
     </div>
 </header>
-<body>
-<h1>메인페이지</h1>
-</body>
+<main>
+    <h1>여행 계획을 세워보세요!</h1>
+    <button type="button" id="mypage-btn" class="btn btn-secondary"
+            onclick="location.href='/getAreaList'">내 여행일정 만들러 가기</button>
+    <button type="button" id="articleList-btn" class="btn btn-secondary"
+            onclick="location.href='/articles'">전체 글 보기</button>
+</main>
+<script src="/js/token.js"></script>
 <script src="/js/header.js"></script>
+</body>
+</html>
