+++ src/main/resources/templates/updateUser.html
@@ -44,7 +44,7 @@
                         </div>
                         <div class="mb-3">
                             <label for="password1" class="form-label text-white">비밀번호</label>
-                            <input type="password" class="form-control" id="password1" placeholder="비밀번호 수정">
+                            <input type="password" class="form-control" id="password1" placeholder="비밀번">
                         </div>
                         <div class="mb-3">
                             <label for="password2" class="form-label text-white">비밀번호 확인</label>
